Searching flickr works, still needs tag filtering on local search

+++ PhotoFinder/DescriptorTools.cs
@@ -209,17 +209,17 @@ namespace PhotoFinder
        // int[] ZrodloCrCDL = new int[3];
        
 
-        public static double[,] PoliczCLD(FileInfo image)
+        public static int[,] PoliczCLD(FileInfo image)
         {
             Bitmap bitmap = (Bitmap)Image.FromFile(image.FullName);
             return PoliczCLD(bitmap);
         }
 
-        public static double[,] PoliczCLD(Bitmap bitmap)
+        public static int[,] PoliczCLD(Bitmap bitmap)
         {
             CLD_Descriptor Mpeg7CLD = new CLD_Descriptor();
 
-            double[,] temp= new double[3,64];
+            int[,] temp= new int[3,64];
 
             Mpeg7CLD.Apply(bitmap);
             for (int h = 0;h <64; h++)
